ShopService: integrated authentication functionalities

+++ ShopService/src/main/java/com/crif/asf/ShopService/controller/CartController.java
@@ -8,13 +8,16 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.crif.asf.ShopService.exception.TokenNotValidException;
 import com.crif.asf.ShopService.model.Cart;
 import com.crif.asf.ShopService.model.CartOrder;
 import com.crif.asf.ShopService.model.Order;
+import com.crif.asf.ShopService.service.AuthService;
 import com.crif.asf.ShopService.service.CartService;
 import com.crif.asf.ShopService.service.OrderService;
 
@@ -27,29 +30,44 @@ public class CartController {
     private CartService cartService;
     @Autowired
     private OrderService orderService;
+    @Autowired
+    private AuthService authService;
 
     // hard-coded user
-    String idUser = "cfed9bb6-44f0-4b4a-be36-0688551f6ea8";
+    // String idUser = "cfed9bb6-44f0-4b4a-be36-0688551f6ea8";
 
     @GetMapping
-    public List<Cart> getCart() {
-	// UUID idUser = UUID.fromString(idUser.toUpperCase());
+    public List<Cart> getCart(@RequestHeader("Authorization") String token) {
+	String idUser = authService.getIdUser(token);
+	if (idUser == null)
+	    throw new TokenNotValidException();
 	return cartService.getCart(idUser);
     }
 
     @PostMapping
-    public void addProductToCart(@RequestBody CartOrder cartOrder) {
+    public void addProductToCart(@RequestBody CartOrder cartOrder,
+	    @RequestHeader("Authorization") String token) {
+	String idUser = authService.getIdUser(token);
+	if (idUser == null)
+	    throw new TokenNotValidException();
 	cartService.addProductToCart(cartOrder, idUser);
     }
 
     @DeleteMapping
     public void removeProductFromCart(
-	    @RequestParam(name = "id_product", required = true) Integer idProduct) {
+	    @RequestParam(name = "id_product", required = true) Integer idProduct,
+	    @RequestHeader("Authorization") String token) {
+	String idUser = authService.getIdUser(token);
+	if (idUser == null)
+	    throw new TokenNotValidException();
 	cartService.removeProductFromCart(idProduct, idUser);
     }
 
     @GetMapping("/buy")
-    public List<Order> buy() {
+    public List<Order> buy(@RequestHeader("Authorization") String token) {
+	String idUser = authService.getIdUser(token);
+	if (idUser == null)
+	    throw new TokenNotValidException();
 	return orderService.buy(idUser);
     }
